really fixed the rounding bug

+++ scores.py
@@ -122,12 +122,12 @@ def get_new_combo():
 def get_penalty(score):
     penalty = score * 0.3
     if penalty > 100:
-        convert_step = penalty / 100
+        convert_step = int(penalty / 100)
     else:
         convert_step = 1
 
     # the following is needed to avoid rounding issues
-    factor = int(penalty) / int(convert_step)
+    factor = penalty / convert_step
     penalty = convert_step * int(factor)
     return -int(penalty), int(convert_step)
